Combine sheets into a single class.
